chore: clippy hygiene sweep — zero non-forbidden warnings
- Remove dead private fields: ParsedTag.meta (markup.rs — metadata lives in
  RawSpan.meta, not ParsedTag), RunningTimer.name (timers.rs — name arg was
  accepted but never stored or read).
- Remove seed_ident_methods!() from CollapsibleContents (private type; id/class
  builder methods were generated but never called).
- Fix unused_mut throughout: css/selectors/parser.rs, many widget constructors.
- Fix field_reassign_with_default in ~15 files: convert post-default field
  assignments to struct-init form (NodeSeed, Style, DispatchOutcome, etc.).
- Fix manual_clamp: key_panel.rs (.min/.max chains → .clamp()), select.rs,
  tabbed_content.rs.
- Fix manual_strip: tabbed_content.rs (starts_with + slice → strip_prefix).
- Fix crate_in_macro_def in delegate.rs: replace all bare crate:: inside
  macro_rules! bodies with $crate::.
- Fix while_let_loop in worker.rs; fix vec_init_then_push in panel.rs.
- Fix map_identity x2 and redundant_locals x2 in layout.rs.
- Fix manual_checked_ops (checked_div/checked_rem) in layout.rs.
- Fix question_mark in routing.rs, container.rs, tabbed_content.rs, mod.rs.
- Fix manual Iterator::find in masked_input.rs.
- Fix empty_line_after_doc_comments in delegate.rs (allow attribute).
- Fix MSRV-1.87 is_multiple_of(100) → % 100 == 0 in progress_bar.rs.
- Add Default impls for OptionList and RadioSet (delegate to new()).
- Add targeted #[allow(clippy::...)] for intentional patterns:
  too_many_arguments (helpers.rs, data_table.rs, tooltip.rs, layout/split.rs,
  content/mod.rs, core.rs), needless_range_loop (layout.rs, layout/grid.rs,
  text.rs), type_complexity (mod.rs, layout/mod.rs), large_enum_variant
  (content/mod.rs), while_let_loop (routing.rs binding-hints loop).
- Removed unused imports (StyleMeta from ~8 files, TryRecvError from worker.rs).
- 25 residual warnings remain in render.rs/scroll_view.rs/scrollbar.rs (owned
  by c11-scrollthumb); all other library code is now warning-free.

+++ src/layout/common.rs
@@ -171,7 +171,7 @@ pub(crate) fn seed_wrapper_subtree_widths(
         let (_, _, bl, br) = border_spacing(&child_style);
         let padding = child_style.effective_padding();
         let outer_inset = margin.left + margin.right;
-        let inner_inset = (bl as u16) + (br as u16) + padding.left + padding.right;
+        let inner_inset = bl + br + padding.left + padding.right;
         let child_outer_w = content_width.saturating_sub(outer_inset).max(1);
         let child_content_w = child_outer_w.saturating_sub(inner_inset).max(1);
         if let Some(node) = tree.get_mut(child) {
@@ -271,8 +271,8 @@ fn horizontal_chrome(
 pub(crate) fn own_box_chrome(style: &Style) -> (u16, u16) {
     let (bt, bb, bl, br) = border_spacing(style);
     let padding = style.effective_padding();
-    let h = (bl as u16) + (br as u16) + padding.left + padding.right;
-    let v = (bt as u16) + (bb as u16) + padding.top + padding.bottom;
+    let h = bl + br + padding.left + padding.right;
+    let v = bt + bb + padding.top + padding.bottom;
     (h, v)
 }
 
@@ -341,10 +341,10 @@ pub(crate) fn extract_child_spec(
     let margin = style.effective_margin();
     let padding = style.effective_padding();
     let (bt, bb, bl, br) = border_spacing(style);
-    let border_top = bt as u16;
-    let border_bottom = bb as u16;
-    let border_left = bl as u16;
-    let border_right = br as u16;
+    let border_top = bt;
+    let border_bottom = bb;
+    let border_left = bl;
+    let border_right = br;
 
     let box_sizing = style.box_sizing.unwrap_or(BoxSizing::BorderBox);
